refactor(nifs): tests of nifs
**Motivation**
In #266 we need to test the protogalaxy folding scheme, for this I want
to reuse circuits from `nifs::vanila`.

Part of #266

**Overview**
- Move test-circuits to `nifs::tests` module
- Move current tests of vanilla to `nifs::vanilla::tests`, instead of `nifs::tests`
- This PR does not add any new code
